new lock-free lifo and fifo
git-svn-id: https://svn.grrrr.org/ext/trunk@1004 4d9ac71a-51e6-0310-8455-cad1006bcd31

+++ source/flqueue.cpp
@@ -155,7 +155,7 @@ static void QWork(bool syslock)
         // qc will be a minimum guaranteed number of present queue elements.
         // On the other hand, if new queue elements are added by the methods called
         // in the loop, these will be sent in the next tick to avoid recursion overflow.
-        int qc = queue.Size();
+        size_t qc = queue.Size();
         if(!qc) break;
 
     #if FLEXT_QMODE == 2
